chore: abstract service implementation
