+++ src/test/java/hng_java_boilerplate/statusPage/service/ApiStatusServiceTest.java
@@ -23,7 +23,7 @@ public class ApiStatusServiceTest {
     private ApiStatusRepository apiStatusRepository;
 
     @InjectMocks
-    private ApiStatusService apiStatusService;
+    private ApiStatusServiceImpl apiStatusService;
 
     @BeforeEach
     public void setUp() {
